+++ src/Query/Delete/DeleteQueryBuilderInterface.php
@@ -3,15 +3,17 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Larva\Query\Delete;
 
-use Closure;
 use WoohooLabs\Larva\Connection\ConnectionInterface;
+use WoohooLabs\Larva\Query\Condition\ConditionBuilderInterface;
 use WoohooLabs\Larva\Query\QueryBuilderInterface;
 
 interface DeleteQueryBuilderInterface extends QueryBuilderInterface
 {
     public function from(string $table): DeleteQueryBuilderInterface;
 
-    public function where(Closure $where): DeleteQueryBuilderInterface;
+    public function where(ConditionBuilderInterface $where): DeleteQueryBuilderInterface;
 
     public function execute(ConnectionInterface $connection): bool;
+
+    public function getQuery(): DeleteQueryInterface;
 }