+++ src/Query/Insert/InsertQueryBuilderInterface.php
@@ -3,9 +3,9 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Larva\Query\Insert;
 
-use Closure;
 use WoohooLabs\Larva\Connection\ConnectionInterface;
 use WoohooLabs\Larva\Query\QueryBuilderInterface;
+use WoohooLabs\Larva\Query\Select\SelectQueryBuilderInterface;
 
 interface InsertQueryBuilderInterface extends QueryBuilderInterface
 {
@@ -17,7 +17,9 @@ interface InsertQueryBuilderInterface extends QueryBuilderInterface
 
     public function multipleValues(array $values): InsertQueryBuilderInterface;
 
-    public function select(Closure $select): InsertQueryBuilderInterface;
+    public function select(SelectQueryBuilderInterface $select): InsertQueryBuilderInterface;
 
     public function execute(ConnectionInterface $connection): bool;
+
+    public function getQuery(): InsertQueryInterface;
 }