+++ src/Query/Select/SelectQueryBuilderInterface.php
@@ -3,9 +3,9 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Larva\Query\Select;
 
-use Closure;
 use Traversable;
 use WoohooLabs\Larva\Connection\ConnectionInterface;
+use WoohooLabs\Larva\Query\Condition\ConditionBuilderInterface;
 use WoohooLabs\Larva\Query\QueryBuilderInterface;
 
 interface SelectQueryBuilderInterface extends QueryBuilderInterface
@@ -20,7 +20,7 @@ interface SelectQueryBuilderInterface extends QueryBuilderInterface
 
     public function from(string $table, string $alias = ""): SelectQueryBuilderInterface;
 
-    public function fromSubquery(Closure $subquery, string $alias): SelectQueryBuilderInterface;
+    public function fromSubquery(SelectQueryBuilderInterface $subquery, string $alias): SelectQueryBuilderInterface;
 
     public function leftJoin(string $table, string $alias = ""): SelectQueryBuilderInterface;
 
@@ -28,15 +28,15 @@ interface SelectQueryBuilderInterface extends QueryBuilderInterface
 
     public function join(string $table, string $alias = "", string $type = ""): SelectQueryBuilderInterface;
 
-    public function on(Closure $on): SelectQueryBuilderInterface;
+    public function on(ConditionBuilderInterface $on): SelectQueryBuilderInterface;
 
-    public function where(Closure $where): SelectQueryBuilderInterface;
+    public function where(ConditionBuilderInterface $where): SelectQueryBuilderInterface;
 
     public function groupBy(string $attribute): SelectQueryBuilderInterface;
 
     public function groupByAttributes(array $attributes): SelectQueryBuilderInterface;
 
-    public function having(Closure $having): SelectQueryBuilderInterface;
+    public function having(ConditionBuilderInterface $having): SelectQueryBuilderInterface;
 
     public function orderBy(string $attribute, string $direction = "ASC"): SelectQueryBuilderInterface;
 
@@ -61,4 +61,6 @@ interface SelectQueryBuilderInterface extends QueryBuilderInterface
     public function fetch(ConnectionInterface $connection): Traversable;
 
     public function fetchColumn(ConnectionInterface $connection): string;
+
+    public function getQuery(): SelectQueryInterface;
 }